Remove Modal Dialog from components
